custom hook for chat room
