+++ src/pages/Chat/components/ChatRoom/ChatRoom.tsx
@@ -1,6 +1,5 @@
 import { Button, Input } from "antd";
 import { SendOutlined } from "@ant-design/icons";
-import { ChangeEvent, useEffect, useState } from "react";
 import {
   Element,
   Header,
@@ -9,25 +8,13 @@ import {
   MessageBlock,
   Notification,
   Warning,
-} from "pages/Chat/ChatRoom/ChatRoom.styles";
-import {
-  useGetMessagesMutation,
-  useJoinRoomMutation,
-  useLeaveRoomMutation,
-  useMessagesQuery,
-  useRoomQuery,
-  useSendMessageMutation,
-} from "services/notifications/setNotificationsAPI";
+} from "pages/Chat/components/ChatRoom/ChatRoom.styles";
 import { formatToStandardDate } from "utils/functions/formatDateFunctions";
-import { useTranslation } from "react-i18next";
 import { CHAT_DATE_FORMAT } from "utils/consts/inputPropsConsts";
 import useChatScroll from "hooks/useChatScroll";
-import {
-  ChatRoomStatusEnum,
-  MessageTypeEnum,
-} from "services/notifications/chatEnums";
-import { useParams } from "react-router-dom";
+import { MessageTypeEnum } from "services/notifications/chatEnums";
 import { UserRoleEnum } from "utils/enums";
+import useChatRoomData from "pages/Chat/hooks/useChatRoomData";
 
 interface IChatRoom {
   userSelfId: number;
@@ -35,79 +22,19 @@ interface IChatRoom {
 }
 
 function ChatRoom({ userSelfId, userRole }: IChatRoom) {
-  const { roomId } = useParams();
+  const {
+    roomId,
+    room,
+    disableInput,
+    t,
+    messages,
+    defineName,
+    onSendHandler,
+    onChangeHandler,
+    text,
+  } = useChatRoomData(userRole);
 
-  const { t } = useTranslation();
-
-  const { data: messages } = useMessagesQuery();
-
-  const { data: room } = useRoomQuery();
-
-  const [sendMessage] = useSendMessageMutation();
-
-  const [getMessages] = useGetMessagesMutation();
-
-  const [joinRoom] = useJoinRoomMutation();
-
-  const [leaveRoom] = useLeaveRoomMutation();
-
-  const [text, setText] = useState<string>("");
-
-  useEffect(() => {
-    if (roomId && +roomId) {
-      joinRoom(+roomId);
-
-      getMessages(+roomId);
-
-      return () => {
-        leaveRoom();
-      };
-    }
-    return () => {
-      leaveRoom();
-    };
-  }, [roomId]);
   const ref = useChatScroll(messages);
-
-  const onChangeHandler = (event: ChangeEvent<HTMLInputElement>) => {
-    setText(event.currentTarget.value);
-  };
-
-  const onSendHandler = () => {
-    if (text.trim() && roomId && +roomId) {
-      sendMessage({ chatRoomId: +roomId, text });
-    }
-    setText("");
-  };
-
-  const defineName = (userID: number): string => {
-    if (room && room.client && room.freelancer) {
-      return userID === room.client.id
-        ? `${room.client.firstName || "User"} ${
-            room.client.lastName || "Client"
-          }`
-        : `${room.freelancer.firstName || "User"} ${
-            room.freelancer.lastName || "Freelancer"
-          }`;
-    }
-    return "User";
-  };
-
-  const disableInput = (): boolean => {
-    if (
-      userRole === UserRoleEnum.CLIENT &&
-      room?.status === ChatRoomStatusEnum.CLIENT_ONLY
-    ) {
-      return false;
-    }
-    if (
-      userRole === UserRoleEnum.FREELANCER &&
-      room?.status === ChatRoomStatusEnum.FREELANCER_ONLY
-    ) {
-      return false;
-    }
-    return room?.status !== ChatRoomStatusEnum.FOR_ALL;
-  };
 
   return (
     <div>
